requirements.txt
absl-py==1.0.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-mnist==0.7
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
